+++ index.js
@@ -1,1 +1,4 @@
-module.exports = require("./src/TreeProgram.js")
+const TreeProgram = require("./src/TreeProgram.js")
+TreeProgram.Tools = require("./src/TreeTools.js")
+
+module.exports = TreeProgram