+++ src/TreeNode.js
@@ -1326,7 +1326,7 @@ class TreeNode extends ImmutableNode {
   }
 
   static getVersion() {
-    return "9.1.0"
+    return "9.2.0"
   }
 }